Module Interval_crlibm

Interval library for OCaml (crlibm version).

This library has the important property that the functions are proved correct (while Interval_intel mostly uses the processor implementation, which is sometimes flawed). It is sometimes faster, sometimes slightly slower than Interval_intel. Some additional functions are also available, thanks to CRlibm.

It is recommended to open this module.

version 1.6

Interval sub-module

type interval = Interval_base.interval = {

low : float;	(*
lower bound, possibly = -∞
*)

high : float;	(*
higher bound, possibly = +∞
*)

}

module I : sig ... end

Interval operations. Locally open this module — using e.g. I.(...) — to redefine classical arithmetic operators for interval arithmetic.

Directed rounding

module type DIRECTED = sig ... end

module RoundDown : DIRECTED

Functions rounding down their results.

module Low = RoundDown

module RoundUp : DIRECTED

Functions rounding up their results.

module High = RoundUp